+++ src/Command/SendCommand.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 namespace DatabaseBackup\Command;
 
 use Cake\Console\Arguments;
+use Cake\Console\CommandFactoryInterface;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
 use DatabaseBackup\Console\Command;
@@ -26,9 +27,20 @@ use Exception;
  * Command to send a backup file via email.
  *
  * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#send
+ * @deprecated 2.13.4' The `SendCommand` is deprecated. Will be removed in a future release
  */
 class SendCommand extends Command
 {
+    public function __construct(?CommandFactoryInterface $factory = null)
+    {
+        parent::__construct($factory);
+
+        deprecationWarning(
+            '2.13.4',
+            'The `SendCommand` is deprecated. Will be removed in a future release'
+        );
+    }
+
     /**
      * @inheritDoc
      */